Renaming ActionButton to SaveButton and updating Bundle size
